module-stat: Fix warning.
Seen when compiling on OSX with `make CC=clang`:

module-stat.c:357:47: warning: conversion specifies type 'unsigned short' but the argument has type 'uint32_t' (aka 'unsigned int') [-Wformat]
                                fprintf(file, "%s,%d,%04hX,%06X,%04hX,%04hX,%d,%d,%ld,%d,%02hX\n",
                                                                      ~~~~^
                                                                      %04u

git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@7308 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ module-stat.c
@@ -356,7 +356,7 @@ void save_stat_to_file_thread(void)
 				//New version:
 				fprintf(file, "%s,%d,%04hX,%06X,%04hX,%04hX,%d,%d,%ld,%d,%02hX\n",
 					rdr->label, stat->rc, stat->caid, stat->prid,
-					stat->srvid, stat->chid, stat->time_avg, stat->ecm_count, stat->last_received, stat->fail_factor, stat->ecmlen);
+					stat->srvid, (uint16_t)stat->chid, stat->time_avg, stat->ecm_count, stat->last_received, stat->fail_factor, stat->ecmlen);
 
 				count++;
 //				if (count % 500 == 0) { //Saving stats is using too much cpu and causes high file load. so we need a break
